+++ CHANGES.md
@@ -1,5 +1,10 @@
 # Changes Logs
 
+## v2.0.0
+
+- Refactored and simpilified the APIs.
+- Fixed the ECDSA signing a invalid P1363 signature.
+
 ## v1.1.0
 
 - Added stream supports.